requirements.txt
django==4.2
djangorestframework==3.14
python-decouple==3.8
paypalrestsdk==1.13.1
stripe==7.0.0
squareup==28.0.0
django-cors-headers